chore: extract sharedHandler
